+++ FluentAvalonia/UI/Controls/CommandBarFlyout/TextCommandBarFlyout.cs
@@ -5,27 +5,12 @@ using Avalonia.Input.Platform;
 using FluentAvalonia.UI.Input;
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using static System.Net.Mime.MediaTypeNames;
 
 namespace FluentAvalonia.UI.Controls
 {
-	public enum TextControlButtons
-	{
-		None = 0x0000,
-		Cut = 0x0001,
-		Copy = 0x0002,
-		Paste = 0x0004,
-		Bold = 0x0008,
-		Italic = 0x0010,
-		Underline = 0x0020,
-		Undo = 0x0040,
-		Redo = 0x0080,
-		SelectAll = 0x0100
-	}
-
+	/// <summary>
+	/// Represents a specialized command bar flyout that contains commands for editing text.
+	/// </summary>
 	public class TextCommandBarFlyout : CommandBarFlyout
 	{
 		public TextCommandBarFlyout()
